junkui/backend/requirements.txt
asgiref==3.2.3
cffi==1.13.2
cryptography==2.8
cycler==0.10.0
decorator==4.4.1
Django==3.0
django-cors-headers==3.2.0
django-extensions==2.2.5
django-werkzeug-debugger-runserver==0.3.1
djangorestframework==3.10.3
h5py==2.10.0
imageio==2.6.1
kiwisolver==1.1.0
matplotlib==3.1.2
networkx==2.4
numpy==1.17.4
Pillow==6.2.1
pycparser==2.19
pyOpenSSL==19.1.0
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
scikit-image==0.16.2
scipy==1.3.3
six==1.13.0
sqlparse==0.3.0
Werkzeug==0.16.0
